requirements.txt
scandir>=1.5
six>=1.10.0
PyYAML==3.12
pytest==3.4.0
